chore: relocates MConnection's bench tests to a separate file (#1187)
As the number of benchmark tests for MConnection grows and their
objectives diverge from those of regular unit tests (their focus shifts
towards performance rather than functional correctness), the benchmark
tests are relocated to a separate file, allowing for enhanced clarity
and maintainability.
